feat(pci-quota): refactor quota regionName
ref: DTCORE-2951
