+++ packages/manager/apps/pci-quota/src/api/data/quota.spec.ts
@@ -10,7 +10,6 @@ describe('Quota API', () => {
       const mockData: Quota[] = [
         {
           region: 'region1',
-          fullRegionName: 'Region 1',
           instance: {
             maxCores: 10,
             maxInstances: 5,
